+++ src/ast_construction/scope_graph_analyzing/mod.rs
@@ -60,7 +60,7 @@ impl<'a> ScopeGraphAnalyzer<'a> {
             .map(|scope_id| (scope_id, ScopeState::Unvisited))
             .collect();
         // Create a total sizes map for all scopes
-        let mut total_sizes: HashMap<ScopeID, usize> = HashMap::new();
+        let mut total_sizes: HashMap<ScopeID, u32> = HashMap::new();
 
         // Analyze the scope graph starting from the global scope
         self.analyze_scope(&global_scope_id, &mut states, &mut total_sizes);
